Handle stack unwinding properly
After reading the source for ART, it turns out that I had interpreted the
method IDs incorrectly - the method ID is overloaded information, but in
addition to enter/exit there was a third option, unwind, that was missing.

This adds unwind support buy dropping the running name stack.
This might result in unexpected graphs if there was a local rewind, but since
the traces don't contain any info on where the stack is unwound to we just
assume that it is wholly unwound each time.